updated servo file

+++ test_servos.py
@@ -7,7 +7,7 @@ import sys
 # kit.servo[0].angle = angle
 
 kit = ServoKit(channels=16)
-throttle_1 = int(sys.argv[1])
-throttle_2 = int(sys.argv[2])
+throttle_1 = float(sys.argv[1])
+throttle_2 = float(sys.argv[2])
 kit.continuous_servo[0].throttle = throttle_1
 kit.continuous_servo[1].throttle = throttle_2
